backend/packages/harness/pyproject.toml
[project]
name = "jobflow-harness"
version = "0.1.0"
description = "JobFlow harness — Career Agent, Middleware Chain, Sub-Agents, Tools"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]
